+++ lib/openapi_first/response.rb
@@ -9,12 +9,13 @@ module OpenapiFirst
   # This is not a direct reflecton of the OpenAPI 3.X response definition, but a combination of
   # status, content type and content schema.
   class Response
-    def initialize(status:, headers:, headers_schema:, content_type:, content_schema:)
+    def initialize(status:, headers:, headers_schema:, content_type:, content_schema:, key:)
       @status = status
       @content_type = content_type
       @content_schema = content_schema
       @headers = headers
       @headers_schema = headers_schema
+      @key = key
       @parser = ResponseParser.new(headers:, content_type:)
       @validator = ResponseValidator.new(self)
     end
@@ -22,7 +23,7 @@ module OpenapiFirst
     # @attr_reader [Integer] status The HTTP status code of the response definition.
     # @attr_reader [String, nil] content_type Content type of this response.
     # @attr_reader [Schema, nil] content_schema the Schema of the response body.
-    attr_reader :status, :content_type, :content_schema, :headers, :headers_schema
+    attr_reader :status, :content_type, :content_schema, :headers, :headers_schema, :key
 
     def validate(response)
       parsed_values = @parser.parse(response)