+++ spec/middlewares/request_validation/request_body_validation_spec.rb
@@ -38,10 +38,6 @@ RSpec.describe 'Request body validation' do
       }
     end
 
-    let(:response_body) do
-      JSON.parse(last_response.body, symbolize_names: true)
-    end
-
     it 'works with stringio' do
       header Rack::CONTENT_TYPE, 'application/json'
       io = StringIO.new(JSON.generate(request_body))
@@ -202,6 +198,32 @@ RSpec.describe 'Request body validation' do
       post path, '{fo},'
 
       expect(last_response.status).to be 400
+    end
+
+    context 'when request body is optional' do
+      let(:path) { '/optional-request-body' }
+
+      it 'accepts a valid request body' do
+        header Rack::CONTENT_TYPE, 'application/json'
+        post path, JSON.generate({ say: 'yes' })
+
+        expect(last_response.status).to be(200), last_response.body
+        expect(last_request.env[OpenapiFirst::REQUEST].parsed_body).to eq 'say' => 'yes'
+      end
+
+      it 'accepts an empty request body' do
+        post path
+
+        expect(last_response.status).to be(200), last_response.body
+        expect(last_request.env[OpenapiFirst::REQUEST].parsed_body).to eq nil
+      end
+
+      it 'returns 400 if request body is invalid' do
+        header Rack::CONTENT_TYPE, 'application/json'
+        post path, JSON.generate({ say: 'no ' })
+
+        expect(last_response.status).to be(400), last_response.body
+      end
     end
 
     context 'with default values' do